requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.2
django-moderation==0.8.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.4
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typed-ast==1.5.4
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
